chore: merge `Mesh` and `InstancedMesh`
