+++ crates/vide/src/api.rs
@@ -1,7 +1,6 @@
 pub mod animation;
 pub mod color;
 pub mod instance;
-pub mod instanced_mesh;
 pub mod mesh;
 pub mod rect;
 pub mod shader;
